Add files via upload
It is a website developed for searching restaurants in your locality(currently added three in my hometown). Here you can search shops and also rate them, so you can know which is most popular and has plenty of services. You can also ask and comment, and search for better dishes accordingly from the menu. Currently looking for more possibilities in the project for further expansion.